EOVOD_MM/requirements/mminstall.txt
mmcls>=0.16.0,<1.0.0
mmcv-full>=1.6.1,<1.7.0
mmdet>=2.19.1,<3.0.0
